Support for Missing Primitive Types: `long`, `double`, `char`
Followup step ensuring that the new types are actually supported in Claro programs. Also updates the stdlib to account for these new types.

+++ src/java/com/claro/intermediate_representation/expressions/numeric/ExponentiateNumericExpr.java
@@ -25,8 +25,7 @@ public class ExponentiateNumericExpr extends NumericExpr {
     ((Expr) this.getChildren().get(0)).assertSupportedExprType(scopedHeap, SUPPORTED_EXPONENTIATE_OPERAND_TYPES);
     ((Expr) this.getChildren().get(1)).assertSupportedExprType(scopedHeap, SUPPORTED_EXPONENTIATE_OPERAND_TYPES);
 
-    // TODO(steving) THIS SHOULD ACTUALLY RETURN DOUBLE.
-    return Types.FLOAT;
+    return Types.DOUBLE;
   }
 
   @Override
